# gen/models/pet.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Pet Pet
//
// 宠物
//
// swagger:model Pet
type Pet struct {

	// age
	// Minimum: > 0
	Age int32 `json:"age,omitempty"`

	// id
	// Minimum: -100
	ID int32 `json:"id,omitempty"`

	// name
	// Pattern: [1|2|3]
	Name string `json:"name,omitempty"`

	// sex
	Sex bool `json:"sex,omitempty"`
}

// Validate validates this pet
func (m *Pet) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateAge(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateName(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Pet) validateAge(formats strfmt.Registry) error {
	if swag.IsZero(m.Age) { // not required
		return nil
	}

	if err := validate.MinimumInt("age", "body", int64(m.Age), 0, true); err != nil {
		return err
	}

	return nil
}

func (m *Pet) validateID(formats strfmt.Registry) error {
	if swag.IsZero(m.ID) { // not required
		return nil
	}

	if err := validate.MinimumInt("id", "body", int64(m.ID), -100, false); err != nil {
		return err
	}

	return nil
}

func (m *Pet) validateName(formats strfmt.Registry) error {
	if swag.IsZero(m.Name) { // not required
		return nil
	}

	if err := validate.Pattern("name", "body", m.Name, `[1|2|3]`); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this pet based on context it is used
func (m *Pet) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Pet) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Pet) UnmarshalBinary(b []byte) error {
	var res Pet
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}